Modularizando componentes de tic-tac-toe
